app: Move ConnectionPopover's AuthorList to it's own subclass
We will add more things to the ConnectionPopverOver so it makes sense to
move the list of authors to a different class.
